Use json checker, not anitya, to find releases (#49)
This project does not actually publish tarballs, and the source being
used is the GitHub archive of the relevant tag. This is served without
any Last-Modified header and with Date set to the current day. This
means that the output of the checker is different every day, since a
different date is used in the metainfo file.

Instead, use the json checker to fetch the release from GitHub via its
API, which includes a timestamp.

Fixes https://github.com/flathub/flatpak-external-data-checker/issues/400